Modified Antique Table Textures, Added Quixel Pipes, Added Several New Materials, Changed Basement
I added an updated texture set for the Antique Table. Several new assets were added and were placed in the scene, mostly the basement. These assets include a slew of models for pipes + their texture set. There are also new wood materials, two of which have opacity maps and the other doesn't.

In regards to the basement, I moved the placement of some shelves and boxes as  well as copied and pasted a box to the other room. I also added wood frames to the wall and pipes to the roof of the basement.